+++ Event/PasteAccountEvent.php
@@ -11,6 +11,7 @@
 
 namespace WBW\Bundle\HaveIBeenPwnedBundle\Event;
 
+use WBW\Bundle\HaveIBeenPwnedBundle\WBWHaveIBeenPwnedEvents;
 use WBW\Library\HaveIBeenPwned\Entity\PasteAccountInterface;
 use WBW\Library\HaveIBeenPwned\Model\Request\PasteAccountRequest;
 use WBW\Library\HaveIBeenPwned\Model\Response\PastesResponse;
@@ -29,7 +30,7 @@ class PasteAccountEvent extends AbstractEvent {
      * @param PasteAccountInterface $entity The paste account.
      */
     public function __construct(PasteAccountInterface $entity) {
-        parent::__construct(HaveIBeenPwnedEvents::PASTE_ACCOUNT, $entity);
+        parent::__construct(WBWHaveIBeenPwnedEvents::PASTE_ACCOUNT, $entity);
     }
 
     /**