+++ Tests/Event/DataClassesEventTest.php
@@ -12,8 +12,8 @@
 namespace WBW\Bundle\HaveIBeenPwnedBundle\Tests\Event;
 
 use WBW\Bundle\HaveIBeenPwnedBundle\Event\DataClassesEvent;
-use WBW\Bundle\HaveIBeenPwnedBundle\Event\HaveIBeenPwnedEvents;
 use WBW\Bundle\HaveIBeenPwnedBundle\Tests\AbstractTestCase;
+use WBW\Bundle\HaveIBeenPwnedBundle\WBWHaveIBeenPwnedEvents;
 use WBW\Library\HaveIBeenPwned\Model\Request\DataClassesRequest;
 use WBW\Library\HaveIBeenPwned\Model\Response\DataClassesResponse;
 
@@ -34,7 +34,7 @@ class DataClassesEventTest extends AbstractTestCase {
 
         $obj = new DataClassesEvent();
 
-        $this->assertEquals(HaveIBeenPwnedEvents::DATA_CLASSES, $obj->getEventName());
+        $this->assertEquals(WBWHaveIBeenPwnedEvents::DATA_CLASSES, $obj->getEventName());
 
         $this->assertNull($obj->getRequest());
         $this->assertNull($obj->getResponse());