+++ Tests/Event/RangeEventTest.php
@@ -11,9 +11,9 @@
 
 namespace WBW\Bundle\HaveIBeenPwnedBundle\Tests\Event;
 
-use WBW\Bundle\HaveIBeenPwnedBundle\Event\HaveIBeenPwnedEvents;
 use WBW\Bundle\HaveIBeenPwnedBundle\Event\RangeEvent;
 use WBW\Bundle\HaveIBeenPwnedBundle\Tests\AbstractTestCase;
+use WBW\Bundle\HaveIBeenPwnedBundle\WBWHaveIBeenPwnedEvents;
 use WBW\Library\HaveIBeenPwned\Model\Request\RangeRequest;
 use WBW\Library\HaveIBeenPwned\Model\Response\RangesResponse;
 
@@ -34,7 +34,7 @@ class RangeEventTest extends AbstractTestCase {
 
         $obj = new RangeEvent($this->range);
 
-        $this->assertEquals(HaveIBeenPwnedEvents::RANGE, $obj->getEventName());
+        $this->assertEquals(WBWHaveIBeenPwnedEvents::RANGE, $obj->getEventName());
 
         $this->assertNull($obj->getRequest());
         $this->assertNull($obj->getResponse());