+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 CHANGELOG
 =========
 
+### [1.3.0](https://github.com/webeweb/haveibeenpwned-bundle/tree/v1.3.0) (2019-07-31)
+
+> IMPORTANT NOTICE: The following classes are deprecated
+> - WBW\Bundle\HaveIBeenPwnedBundle\Event\HaveIBeenPwnedEvents
+
+- Add HaveIBeenPwned events
+
 ### [1.2.1](https://github.com/webeweb/haveibeenpwned-bundle/tree/v1.2.1) (2019-06-08)
 
 - Fix deprecated root() call on TreeBuilder